feat!: Different server/client flavors use Swift traits
This required bumping to Swift 6.1
