+++ .swiftformat
@@ -1,0 +1,1 @@
+--ifdef no-indent
