+++ Sources/HaystackClient/ClientDarwin/Client+URLSession.swift
@@ -1,25 +1,22 @@
-import AsyncHTTPClient
+#if ClientDarwin
 import Foundation
-import HaystackClient
-import NIO
 
 public extension Client {
-    /// Create a Haystack API Client that uses a NIO-based HTTP client.
+    /// Create a Haystack API Client that uses a `URLSession` from `Foundation` that
+    /// is only available on Darwin platforms.
     ///
     /// - Parameters:
     ///   - baseUrl: The URL of the Haystack API server
     ///   - username: The username to authenticate with
     ///   - password: The password to authenticate with
     ///   - format: The transfer data format. Defaults to `zinc` to reduce data transfer.
-    ///   - eventLoopGroup: The event loop group on which to create request callbacks
     convenience init(
         baseUrl: String,
         username: String,
         password: String,
-        format: DataFormat = .zinc,
-        httpClient: HTTPClient
+        format: DataFormat = .zinc
     ) throws {
-        let fetcher = httpClient.haystackFetcher()
+        let fetcher = URLSessionFetcher()
         try self.init(
             baseUrl: baseUrl,
             username: username,
@@ -29,3 +26,4 @@ public extension Client {
         )
     }
 }
+#endif
