--- Sources/HaystackClientNIO/Exports.swift
@@ -1,1 +1,0 @@
-@_exported import HaystackClient
