+++ Sources/HaystackServer/ServerVapor/HaystackRouteCollection.swift
@@ -1,3 +1,4 @@
+#if ServerVapor
 import Haystack
 import Vapor
 
@@ -551,3 +552,4 @@ public struct HaystackRouteCollection: RouteCollection {
         }
     }
 }
+#endif
